Module Setup
This is to make the harvest importable and usable as a python module for other python projects. CLI has been broken out into its own script, and utilizes setuptools console_scripts entrypoint.
