+++ .gitignore
@@ -3,6 +3,9 @@ __pycache__/
 *.py[cod]
 *$py.class
 
+#PyCharm
+.idea/
+
 # C extensions
 *.so
 
@@ -68,3 +71,4 @@ RemoteSystemsTempFiles
 .metadata
 .settings
 *.prefs
+venv/
